Switch from REST API to GraphQL API
Using the REST API on a repo with `n` branches would result in at least
`n+1` calls to query the author and commit date for each branch. Using
the GraphQL API, the required data can returned in a much smaller number
of calls. In most cases this should be an order of magnitude fewer
calls.